+++ azure-application-insights-spring-boot-starter/src/main/java/com/microsoft/applicationinsights/autoconfigure/ApplicationInsightsTelemetryAutoConfiguration.java
@@ -183,7 +183,8 @@ public class ApplicationInsightsTelemetryAutoConfiguration {
             LocalForwarder lf = applicationInsightsProperties.getChannel().getLocalForwarder();
 
             // only the endpoint address is picked up. All the other values are fake and just for future use if needed.
-            return new LocalForwarderTelemetryChannel(lf.getEndpointAddress(), false, 0 ,0);
+            return new LocalForwarderTelemetryChannel(lf.getEndpointAddress(), false, lf.getMaxTelemetryBufferCapacity() ,
+                lf.getFlushIntervalInSeconds());
         }
 
         InProcess inProcess = applicationInsightsProperties.getChannel().getInProcess();