+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # Version 2.2.0
 - Introduces SpringBoot Starter 1.1.0-BETA
 - Starter now respects autoconfiguration for Micrometer metrics.
-- Starter adds autoconfiguration for Local Forwarder gRPC Telemetry Channel. (Please look at readme for details on configuration.)
+- Starter adds autoconfiguration for Local Forwarder Telemetry Channel. (Please look at readme for details on configuration.)
 - Fix [#712](https://github.com/Microsoft/ApplicationInsights-Java/issues/712) the thread shutdown issue in SpringBoot Starter by registering `ApplicationInsightsServletContextListener`.
 - SpringBoot Starter now supports reading iKey using all the variable names as core sdk.
 - Starter would no longer support relaxed binding of ikey property due to complex conditional need and backport problems with RelaxedBinder from Boot 2 to 1.5.x.